WIP fetch CAS blobs from IPFS
The final commit that might be proposed to upstream would look more
like a config entry for a (list of?) URI templates where to try to
fetch CAS blobs, and we'd just configure it to point to the IPFS API.

One con of the approach here is that the image needs to be stored in a
single IPFS block (`ipfs block put`, not `ipfs add`). If that is not
done, the CAS hashes don't match and somewhere there'd need to be
a mapping layer; at which point we might as well go with ipns.

The `ipfs block put` needs to be done with a patched ipfs, that uses
SHA-512 truncated to 256 bits. In github.com/ipfs/go-ipfs/util/util.go,

     // Hash is the global IPFS hash function. uses multihash SHA2_256, 256 bits
     func Hash(data []byte) mh.Multihash {
    -       h, err := mh.Sum(data, mh.SHA2_256, -1)
    +       h, err := mh.Sum(data, mh.SHA2_512, 32)

It's ok to build that special version only to run a local mode `ipfs
block put` with it, and then start up an unpatched `ipfs daemon`.